requirements.txt
Flask>=3.0
flask-cors>=4.0
openai>=1.0
gunicorn>=21.2
pillow
ollama